+++ src/hooks/useStoryEvents.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const useStoryEvents = () => {
+  const [events, setEvents] = React.useState([])
+
+  return [events];
+}
